+++ internal/v1/creditcards/services/delete_creditcard.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"financialcontrol/internal/constants"
 	e "financialcontrol/internal/models/errors"
 	"net/http"
 
@@ -14,7 +15,15 @@ func (c CreditCardsService) Delete(ctx *gin.Context) (int, []e.ApiError) {
 		return status, err
 	}
 
-	// TODO: check if creditcard has transactions associated
+	hasTransactions, err := c.repository.HasTransactionsByCreditCard(ctx, creditcard.ID)
+
+	if len(err) > 0 {
+		return http.StatusInternalServerError, err
+	}
+
+	if hasTransactions {
+		return http.StatusBadRequest, []e.ApiError{e.CustomError{Message: constants.CreditcardCannotBeDeletedMsg}}
+	}
 
 	err = c.repository.Delete(ctx, creditcard.ID)
